@@ -1,6 +1,9 @@
 {
   "name": "kbn_sunburst_vis",
-  "version": "1.0.0",
+  "version": "6.2.2",
+  "kibana": {
+    "version": "6.2.2"
+  },
   "authors": [
     "Juan Carniglia <juan.carniglia@gmail.com>"
   ],
@@ -18,26 +21,17 @@
   },
   "devDependencies": {
     "babel-core": "^6.5.2",
-    "babel-eslint": "^4.1.8",
+    "babel-eslint": "^7.2.1",
     "babel-loader": "^6.2.3",
-    "babel-preset-es2015": "^6.5.0",
-    "babel-register": "^6.5.2",
-    "bluebird": "^3.3.1",
-    "bower": "^1.6.9",
-    "del": "^2.2.0",
-    "eslint": "^2.1.0",
-    "eslint-plugin-mocha": "^2.0.0",
-    "gulp": "^3.9.1",
-    "gulp-eslint": "^2.0.0",
-    "gulp-gzip": "^1.2.0",
-    "gulp-tar": "^1.8.0",
-    "gulp-util": "^3.0.7",
-    "gulp-zip": "3.1.0",
+    "bluebird": "^3.5.0",
+    "eslint": "^3.18.0",
+    "eslint-plugin-mocha": "^4.9.0",
     "husky": "^0.10.2",
     "lodash": "^4.4.0",
     "minimist": "1.2.0",
     "mkdirp": "^0.5.1",
-    "rimraf": "^2.4.4",
-    "rsync": "^0.4.0"
+    "rimraf": "^2.6.1",
+    "rsync": "^0.6.1",
+    "webpack": "^2.2.0-rc"
   }
 }
